requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
gunicorn==21.2.0
python-jose[cryptography]==3.3.0  # для JWT токенов
passlib[bcrypt]==1.7.4  # для хеширования паролей
psycopg2-binary>=2.9.9
python-dotenv==1.0.0